backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
openpyxl
minio
google-genai
python-multipart
ortools
requests
numpy
networkx
